// generated/src/app/api-models/prem-bill-sel-groups-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PremBillSelGroupsWork {
  seqGpbilId: number;
  seqGroupId: number;
  groupId: string;
  shortName: string;
  seqParentId: number;
  customerType: string;
  customerId: string;
  groupName1: string;
  groupName2: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zipCode: string;
  groupType: string;
  prorationMethod: string;
  billingLevel: string;
  billingCycle: string;
  billingBatch: string;
  noOfRetroMonths: number;
  billedThroughDate: Date;
  billingDatetime: Date;
  sortOption: string;
  retroRatesFlag: string;
  country: string;
}